+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Experimental
+
 ## [1.137.0] - Not released
+### Changed
+- Totally new design of post attachments, including larger previews and denser
+  gallery layout.
+- Support for video attachments, asynchronous media processing and the V4 server
+  API.
+### Fixed
+- File uploader now properly displays the upload progress.
 
 ## [1.136.3] - 2025-01-05
 ### Changed